Refactor multi-object tracker shape model and scoring system
- Replaced the existing shapes.cpp with shapes_iou.cpp and shapes_transform.cpp to enhance the shape handling capabilities.
- Updated CMakeLists.txt to include new shape model files.
- Modified scoring algorithms (bev_assignment_scoring, polar_assignment_scoring, redundancy_check) to utilize the new shapes_iou.hpp for improved intersection over union calculations.
- Implemented new functions in shapes_iou.cpp for 1D and 2D IoU calculations, generalized IoU, and precision-recall metrics.
- Introduced shapes_transform.cpp for transforming polygon footprints and computing bounding boxes.
- Updated various shape model implementations (pedestrian, static, vehicle) to use shapes_transform.hpp instead of the deprecated shapes.hpp.
- Removed unnecessary includes of shapes.hpp across multiple files to streamline dependencies.
