+++ perception/autoware_multi_object_tracker/include/autoware/multi_object_tracker/object_model/shapes_iou.hpp
@@ -1,0 +1,49 @@
+// Copyright 2024 TIER IV, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef AUTOWARE__MULTI_OBJECT_TRACKER__OBJECT_MODEL__SHAPES_IOU_HPP_
+#define AUTOWARE__MULTI_OBJECT_TRACKER__OBJECT_MODEL__SHAPES_IOU_HPP_
+
+#include "autoware/multi_object_tracker/types.hpp"
+
+#include <utility>
+
+namespace autoware::multi_object_tracker
+{
+namespace shapes
+{
+
+double get1dIoU(
+  const types::DynamicObject & source_object, const types::DynamicObject & target_object);
+
+double get2dIoU(
+  const types::DynamicObject & source_object, const types::DynamicObject & target_object,
+  const double min_union_area = 0.01);
+
+double get2dGeneralizedIoU(
+  const types::DynamicObject & source_object, const types::DynamicObject & target_object);
+
+bool get2dPrecisionRecallGIoU(
+  const types::DynamicObject & source_object, const types::DynamicObject & target_object,
+  double & precision, double & recall, double & generalized_iou);
+
+std::pair<double, double> getObjectZRange(const types::DynamicObject & object);
+
+double get3dGeneralizedIoU(
+  const types::DynamicObject & source_object, const types::DynamicObject & target_object);
+
+}  // namespace shapes
+}  // namespace autoware::multi_object_tracker
+
+#endif  // AUTOWARE__MULTI_OBJECT_TRACKER__OBJECT_MODEL__SHAPES_IOU_HPP_
