+++ perception/autoware_multi_object_tracker/include/autoware/multi_object_tracker/object_model/shapes_transform.hpp
@@ -12,34 +12,19 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef AUTOWARE__MULTI_OBJECT_TRACKER__OBJECT_MODEL__SHAPES_HPP_
-#define AUTOWARE__MULTI_OBJECT_TRACKER__OBJECT_MODEL__SHAPES_HPP_
+#ifndef AUTOWARE__MULTI_OBJECT_TRACKER__OBJECT_MODEL__SHAPES_TRANSFORM_HPP_
+#define AUTOWARE__MULTI_OBJECT_TRACKER__OBJECT_MODEL__SHAPES_TRANSFORM_HPP_
 
 #include "autoware/multi_object_tracker/types.hpp"
 
 #include <geometry_msgs/msg/point.hpp>
 
 #include <optional>
-#include <utility>
 
 namespace autoware::multi_object_tracker
 {
 namespace shapes
 {
-
-double get1dIoU(
-  const types::DynamicObject & source_object, const types::DynamicObject & target_object);
-
-double get2dIoU(
-  const types::DynamicObject & source_object, const types::DynamicObject & target_object,
-  const double min_union_area = 0.01);
-
-double get2dGeneralizedIoU(
-  const types::DynamicObject & source_object, const types::DynamicObject & target_object);
-
-bool get2dPrecisionRecallGIoU(
-  const types::DynamicObject & source_object, const types::DynamicObject & target_object,
-  double & precision, double & recall, double & generalized_iou);
 
 bool convertConvexHullToBoundingBox(
   const types::DynamicObject & input_object, types::DynamicObject & output_object,
@@ -47,11 +32,6 @@ bool convertConvexHullToBoundingBox(
 
 std::optional<types::DynamicObject> alignClusterToOrientation(
   const types::DynamicObject & cluster, double target_yaw);
-
-std::pair<double, double> getObjectZRange(const types::DynamicObject & object);
-
-double get3dGeneralizedIoU(
-  const types::DynamicObject & source_object, const types::DynamicObject & target_object);
 
 // Transform polygon footprint points from src_pose's local frame into dst_pose's local frame.
 // Equivalent to: p_dst = R_dst^T * (R_src * p_src + t_src - t_dst)
@@ -68,4 +48,4 @@ geometry_msgs::msg::Polygon unionFootprints(
 }  // namespace shapes
 }  // namespace autoware::multi_object_tracker
 
-#endif  // AUTOWARE__MULTI_OBJECT_TRACKER__OBJECT_MODEL__SHAPES_HPP_
+#endif  // AUTOWARE__MULTI_OBJECT_TRACKER__OBJECT_MODEL__SHAPES_TRANSFORM_HPP_
