+++ perception/autoware_multi_object_tracker/lib/object_model/shapes_transform.cpp
@@ -12,9 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/multi_object_tracker/object_model/shapes.hpp"
+#include "autoware/multi_object_tracker/object_model/shapes_transform.hpp"
 
-#include <Eigen/Geometry>
 #include <autoware_utils_geometry/boost_geometry.hpp>
 #include <autoware_utils_geometry/boost_polygon_utils.hpp>
 #include <tf2/utils.hpp>
@@ -24,18 +23,13 @@
 
 #include <boost/geometry.hpp>
 
-#include <algorithm>
 #include <cmath>
 #include <limits>
-#include <string>
-#include <utility>
+#include <optional>
 #include <vector>
 
 namespace
 {
-constexpr double MIN_AREA = 1e-6;
-constexpr double INVALID_SCORE = -1.0;
-
 struct OrientedExtent
 {
   double min_along, max_along, min_lat, max_lat;
@@ -59,146 +53,12 @@ inline OrientedExtent computeOrientedExtent(
   }
   return ext;
 }
-double getSumArea(const std::vector<autoware_utils_geometry::Polygon2d> & polygons)
-{
-  return std::accumulate(
-    polygons.begin(), polygons.end(), 0.0, [](double acc, autoware_utils_geometry::Polygon2d p) {
-      return acc + boost::geometry::area(p);
-    });
-}
-
-double getIntersectionArea(
-  const autoware_utils_geometry::Polygon2d & source_polygon,
-  const autoware_utils_geometry::Polygon2d & target_polygon)
-{
-  std::vector<autoware_utils_geometry::Polygon2d> intersection_polygons;
-  boost::geometry::intersection(source_polygon, target_polygon, intersection_polygons);
-  return getSumArea(intersection_polygons);
-}
-
-double getUnionArea(
-  const autoware_utils_geometry::Polygon2d & source_polygon,
-  const autoware_utils_geometry::Polygon2d & target_polygon)
-{
-  std::vector<autoware_utils_geometry::Polygon2d> union_polygons;
-  boost::geometry::union_(source_polygon, target_polygon, union_polygons);
-  return getSumArea(union_polygons);
-}
-
-double getConvexShapeArea(
-  const autoware_utils_geometry::Polygon2d & source_polygon,
-  const autoware_utils_geometry::Polygon2d & target_polygon)
-{
-  boost::geometry::model::multi_polygon<autoware_utils_geometry::Polygon2d> union_polygons;
-  boost::geometry::union_(source_polygon, target_polygon, union_polygons);
-
-  autoware_utils_geometry::Polygon2d hull;
-  boost::geometry::convex_hull(union_polygons, hull);
-  return boost::geometry::area(hull);
-}
 }  // namespace
 
 namespace autoware::multi_object_tracker
 {
 namespace shapes
 {
-
-double get1dIoU(
-  const types::DynamicObject & source_object, const types::DynamicObject & target_object)
-{
-  constexpr double min_union_length = 0.1;  // As 0.01 used in 2dIoU, use 0.1 here
-  constexpr double min_length = 1e-3;       // As 1e-6 used in 2dIoU, use 1e-3 here
-  // Compute radii from dimensions (use max of x and y as diameter)
-  const double r_src =
-    std::max(source_object.shape.dimensions.x, source_object.shape.dimensions.y) * 0.5;
-  const double r_tgt =
-    std::max(target_object.shape.dimensions.x, target_object.shape.dimensions.y) * 0.5;
-  // if radius is smaller than the minimum length, return 0.0
-  if (r_src < min_length || r_tgt < min_length) return 0.0;
-  // Ensure r1 is the larger radius
-  const double r1 = std::max(r_tgt, r_src);
-  const double r2 = std::min(r_tgt, r_src);
-  const auto dx = source_object.pose.position.x - target_object.pose.position.x;
-  const auto dy = source_object.pose.position.y - target_object.pose.position.y;
-  // distance between centers
-  const auto dist = std::sqrt(dx * dx + dy * dy);
-  // if distance is larger than the sum of radius, return 0.0
-  if (dist > r1 + r2 - min_length) return 0.0;
-  // if distance is smaller than the difference of radius, return the ratio of the smaller radius to
-  // the larger radius
-  // Square used to mimic area ratio behavior as a rough 2D approximation
-  if (dist < r1 - r2) return (r2 * r2) / (r1 * r1);
-  // if distance is between the difference and the sum of radii, return the ratio of the
-  // intersection length to the union length
-  if (r1 + r2 + dist < min_union_length) return 0.0;
-  const double intersection_length = r1 + r2 - dist;
-  const double iou = intersection_length * r2 / (r1 * r1) * 0.5;
-  return iou;
-}
-
-double get2dIoU(
-  const types::DynamicObject & source_object, const types::DynamicObject & target_object,
-  const double min_union_area)
-{
-  const auto source_polygon =
-    autoware_utils_geometry::to_polygon2d(source_object.pose, source_object.shape);
-  if (boost::geometry::area(source_polygon) < MIN_AREA) return 0.0;
-  const auto target_polygon =
-    autoware_utils_geometry::to_polygon2d(target_object.pose, target_object.shape);
-  if (boost::geometry::area(target_polygon) < MIN_AREA) return 0.0;
-
-  const double intersection_area = getIntersectionArea(source_polygon, target_polygon);
-  if (intersection_area < MIN_AREA) return 0.0;
-  const double union_area = getUnionArea(source_polygon, target_polygon);
-
-  const double iou =
-    union_area < min_union_area ? 0.0 : std::min(1.0, intersection_area / union_area);
-  return iou;
-}
-
-double get2dGeneralizedIoU(
-  const types::DynamicObject & source_object, const types::DynamicObject & target_object)
-{
-  const auto source_polygon =
-    autoware_utils_geometry::to_polygon2d(source_object.pose, source_object.shape);
-  const double source_area = boost::geometry::area(source_polygon);
-  const auto target_polygon =
-    autoware_utils_geometry::to_polygon2d(target_object.pose, target_object.shape);
-  const double target_area = boost::geometry::area(target_polygon);
-  if (source_area < MIN_AREA && target_area < MIN_AREA) return -1.0;
-
-  const double intersection_area = getIntersectionArea(source_polygon, target_polygon);
-  const double union_area = getUnionArea(source_polygon, target_polygon);
-  const double iou = union_area < 0.01 ? 0.0 : std::min(1.0, intersection_area / union_area);
-  const double convex_shape_area = getConvexShapeArea(source_polygon, target_polygon);
-
-  return iou - (convex_shape_area - union_area) / convex_shape_area;
-}
-
-bool get2dPrecisionRecallGIoU(
-  const types::DynamicObject & source_object, const types::DynamicObject & target_object,
-  double & precision, double & recall, double & generalized_iou)
-{
-  const auto source_polygon =
-    autoware_utils_geometry::to_polygon2d(source_object.pose, source_object.shape);
-  const double source_area = boost::geometry::area(source_polygon);
-  if (source_area < MIN_AREA) return false;
-  const auto target_polygon =
-    autoware_utils_geometry::to_polygon2d(target_object.pose, target_object.shape);
-  const double target_area = boost::geometry::area(target_polygon);
-  if (target_area < MIN_AREA) return false;
-
-  const double intersection_area = getIntersectionArea(source_polygon, target_polygon);
-  const double union_area = getUnionArea(source_polygon, target_polygon);
-  const double convex_shape_area = getConvexShapeArea(source_polygon, target_polygon);
-  const double iou = union_area < 0.01 ? 0.0 : std::min(1.0, intersection_area / union_area);
-
-  precision = source_area < MIN_AREA ? 0.0 : std::min(1.0, intersection_area / source_area);
-  recall = source_area < MIN_AREA ? 0.0 : std::min(1.0, intersection_area / target_area);
-  generalized_iou = iou - (convex_shape_area - union_area) / convex_shape_area;
-
-  return true;
-}
 
 bool convertConvexHullToBoundingBox(
   const types::DynamicObject & input_object, types::DynamicObject & output_object,
@@ -356,47 +216,6 @@ std::optional<types::DynamicObject> alignClusterToOrientation(
   aligned.shape.dimensions.y = ext.max_lat - ext.min_lat;
 
   return aligned;
-}
-
-std::pair<double, double> getObjectZRange(const types::DynamicObject & object)
-{
-  const double center_z = object.pose.position.z;
-  const double height = object.shape.dimensions.z;
-  const double min_z = center_z - height / 2.0;
-  const double max_z = center_z + height / 2.0;
-  return {min_z, max_z};
-}
-
-double get3dGeneralizedIoU(
-  const types::DynamicObject & source_object, const types::DynamicObject & target_object)
-{
-  const auto source_polygon =
-    autoware_utils_geometry::to_polygon2d(source_object.pose, source_object.shape);
-  if (boost::geometry::area(source_polygon) < MIN_AREA) return INVALID_SCORE;
-  const auto target_polygon =
-    autoware_utils_geometry::to_polygon2d(target_object.pose, target_object.shape);
-  if (boost::geometry::area(target_polygon) < MIN_AREA) return INVALID_SCORE;
-
-  const double union_area = getUnionArea(source_polygon, target_polygon);
-  if (union_area < MIN_AREA) return INVALID_SCORE;
-
-  const double intersection_area = getIntersectionArea(source_polygon, target_polygon);
-  const double convex_area = getConvexShapeArea(source_polygon, target_polygon);
-
-  const auto [z_min_src, z_max_src] = getObjectZRange(source_object);
-  const auto [z_min_tgt, z_max_tgt] = getObjectZRange(target_object);
-
-  const double height_overlap =
-    std::max(0.0, std::min(z_max_src, z_max_tgt) - std::max(z_min_src, z_min_tgt));
-
-  if (height_overlap <= 0.0) return INVALID_SCORE;
-
-  const double total_height = std::max(z_max_src, z_max_tgt) - std::min(z_min_src, z_min_tgt);
-
-  const double iou =
-    std::clamp((intersection_area * height_overlap) / (union_area * total_height), 0.0, 1.0);
-
-  return iou - (convex_area - union_area) / convex_area;
 }
 
 geometry_msgs::msg::Polygon unionFootprints(
