+++ perception/autoware_multi_object_tracker/lib/tracker/shape_model/vehicle_shape_model.cpp
@@ -14,7 +14,7 @@
 
 #include "autoware/multi_object_tracker/tracker/shape_model/vehicle_shape_model.hpp"
 
-#include "autoware/multi_object_tracker/object_model/shapes.hpp"
+#include "autoware/multi_object_tracker/object_model/shapes_transform.hpp"
 
 #include <algorithm>
