+++ perception/autoware_multi_object_tracker/lib/types.cpp
@@ -14,7 +14,6 @@
 
 #include "autoware/multi_object_tracker/types.hpp"
 
-#include "autoware/multi_object_tracker/object_model/shapes.hpp"
 #include "autoware/multi_object_tracker/object_model/uuid.hpp"
 
 #include <cmath>
