+++ perception/autoware_multi_object_tracker/src/multi_object_tracker_node.cpp
@@ -16,7 +16,6 @@
 
 #include "multi_object_tracker_node.hpp"
 
-#include "autoware/multi_object_tracker/object_model/shapes.hpp"
 #include "autoware/multi_object_tracker/types.hpp"
 #include "autoware/multi_object_tracker/uncertainty/uncertainty_processor.hpp"
