+++ perception/autoware_multi_object_tracker/test/test_tracker_overlap_manager.cpp
@@ -15,7 +15,7 @@
 #include "autoware/multi_object_tracker/association/scoring/redundancy_check.hpp"
 #include "autoware/multi_object_tracker/association/tracker_overlap_manager.hpp"
 #include "autoware/multi_object_tracker/object_model/object_model.hpp"
-#include "autoware/multi_object_tracker/object_model/shapes.hpp"
+#include "autoware/multi_object_tracker/object_model/shapes_iou.hpp"
 #include "autoware/multi_object_tracker/tracker/trackers/polygon_tracker.hpp"
 #include "autoware/multi_object_tracker/tracker/trackers/vehicle_tracker.hpp"
 #include "autoware/multi_object_tracker/types.hpp"
